added archlinux-specific checks to prevent code execution in OSX

+++ install/packer.sh
@@ -1,20 +1,24 @@
 #!/usr/bin/env bash
 
-if test ! $(which packer); then
-	echo "Installing packer"
+if [[ $CURRENT_OS == 'Linux' ]]; then
+	if test ! $(which packer); then
+		echo "Installing packer"
 
-	# packer dependencies
-	sudo pacman -S git wget fakeroot jshon expac base-devel
+		# packer dependencies
+		sudo pacman -S git wget fakeroot jshon expac base-devel
 
-	pushd
-	cd /tmp
-	mkdir packer
-	cd packer
-	wget https://aur.archlinux.org/cgit/aur.git/plain/PKGBUILD?h=packer -O PKGBUILD
-	makepkg
-	sudo pacman -U $(ls | grep packer-)
-	popd
-	rm -rf /tmp/packer
+		pushd
+		cd /tmp
+		mkdir packer
+		cd packer
+		wget https://aur.archlinux.org/cgit/aur.git/plain/PKGBUILD?h=packer -O PKGBUILD
+		makepkg
+		sudo pacman -U $(ls | grep packer-)
+		popd
+		rm -rf /tmp/packer
+	else
+		printf "packer is already installed."
+	fi
 else
-	printf "packer is already installed."
+	printf "Aborting, Archlinux specific"
 fi
